requirements.txt
Django==4.2
django-cors-headers==3.8.0
djangorestframework==3.12.4
mysqlclient==2.0.3
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0